Moved hyperopt params to strategy file
