+++ freqtrade/strategy/strategy.py
@@ -8,6 +8,7 @@ import sys
 import logging
 import importlib
 from collections import OrderedDict
+from typing import Dict, Any, Callable
 
 from pandas import DataFrame
 from freqtrade.strategy.interface import IStrategy
@@ -178,3 +179,22 @@ class Strategy(object):
         :return: DataFrame with buy column
         """
         return self.custom_strategy.populate_sell_trend(dataframe)
+
+    def indicator_space(self) -> Dict[str, Any]:
+        """
+        Define your Hyperopt space for searching strategy parameters
+        """
+        return self.custom_strategy.indicator_space()
+        
+    def buy_strategy_generator(self, params: Dict[str, Any]) -> Callable:
+        """
+        Define the buy strategy parameters to be used by hyperopt
+        """
+        return self.custom_strategy.buy_strategy_generator(params)
+        
+    def roi_space(self) -> Dict[str, Any]:
+        return self.custom_strategy.roi_space()
+        
+    def stoploss_space(self) -> Dict[str, Any]:
+        return self.custom_strategy.stoploss_space()
+
